+++ react/packages/@porsche/ui-kit-react-docs/src/examples/molecules/Select/index.js
@@ -1,18 +1,13 @@
-import React from "react"
 import Single from "./Single"
 import Multi from "./Multi"
 import OptionGroups from "./OptionGroups"
 import EdgeCases from "./EdgeCases"
 
-const SelectExamples = () => {
-    return (
-        <div>
-            <Single />
-            <Multi />
-            <OptionGroups />
-            <EdgeCases />
-        </div>
-    )
-}
+import { Select } from "@porsche/ui-kit-react"
+import { storiesOf, TYPE } from "src/app/stories"
 
-export default SelectExamples
+storiesOf(TYPE.MOLECULE, Select, [], module)
+    .add(Single)
+    .add(Multi)
+    .add(OptionGroups)
+    .add(EdgeCases)