+++ react/packages/@porsche/ui-kit-react-docs/src/examples/organisms/Modal/Content/ModalExampleContentTitleText.js
@@ -2,7 +2,7 @@ import React from "react"
 import { Modal, Button } from "@porsche/ui-kit-react"
 import lorem from "lorem-ipsum"
 
-class ModalExampleContentTitleText extends React.Component {
+class ModalExampleContentTitleText extends React.PureComponent {
     constructor(props) {
         super(props)