+++ react/packages/@porsche/ui-kit-react-docs/src/examples/organisms/Toast/ToastExampleSimple.js
@@ -1,0 +1,6 @@
+import React from "react"
+import { Toast } from "@porsche/ui-kit-react"
+
+const ToastExampleSimple = () => <Toast message="Hi, I'm a Toast!" onClick={() => {}} />
+
+export default ToastExampleSimple