+++ react/packages/@porsche/ui-kit-react/src/components/molecules/Loader/Loader.tsx
@@ -1,9 +1,13 @@
 import * as React from "react"
 import cx from "classnames"
 
-import { ComponentMeta, MetaCategorizable } from "../../../types/MetaCategorizable"
-import { getElementType, META, prefix } from "../../../lib"
+import { getElementType, prefix } from "../../../lib"
 import { ClassNameProp, ComponentProp } from "../../../lib/props"
+import { LoaderMask } from "./LoaderMask"
+
+export interface Loader extends React.StatelessComponent<LoaderProps> {
+    Mask: typeof LoaderMask
+}
 
 export interface LoaderProps extends ClassNameProp, ComponentProp {
     /**
@@ -17,12 +21,7 @@ export interface LoaderProps extends ClassNameProp, ComponentProp {
     size?: "default" | "small"
 }
 
-const _meta: ComponentMeta = {
-    name: "Loader",
-    type: META.TYPES.MOLECULE
-}
-
-const _Loader: React.StatelessComponent<LoaderProps> & Partial<MetaCategorizable> = (props) => {
+const _Loader: React.StatelessComponent<LoaderProps> & Partial<Loader> = (props) => {
     const { as, className, children, inverted, size, ...rest } = props
 
     const ElementType = getElementType(as, "div")
@@ -46,10 +45,10 @@ const _Loader: React.StatelessComponent<LoaderProps> & Partial<MetaCategorizable
     )
 }
 
-_Loader._meta = _meta
+_Loader.Mask = LoaderMask
 
 /**
  * A loader component to show loading states inside single components or across entire modules / pages.
  * @see Button
  */
-export const Loader = _Loader as React.StatelessComponent<LoaderProps>
+export const Loader = _Loader as Loader