+++ react/packages/@porsche/ui-kit-react/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.2.0] - 2018-10-23
+- added `Toast`and `Toast.Manager` component to display little notifications or messages on the bottom left.
+- added `loading` prop to Modal to show a loading indicator above the content but below the close button
+- added `Loader.Mask` component to complement Loader
+
 ## [0.1.33] - 2018-10-15
 - added bold font-weight to `text` component
 - added `autofocus` property to `input` and `textarea` component